Major Working Refactor.
